requirements.txt
pandas
pillow
click
setuptools
numpy
rpy2
pygeoapi
opencdms@git+https://github.com/opencdms/pyopencdms.git#egg=opencdms
